refactor(agent): step 改为 class、每步独立成 *-step.ts 文件，describe/review 拆为两步
按可维护性诉求把 step 落成 OOP 抽象 + 每步一文件（行为除 describe/review 拆分外不变）：
- steps/context：StepHandler 接口改为抽象基类 Step<Ctx, R=void>，子类实现 run 方法（去内联 run 闭包）。
- steps/review/：describe-step / review-step / judge-step / asks-step / summary-step 各一 class 文件，
  + shared.ts（ReviewStepCtx / 判读·总结提示词与解析）+ index.ts（REVIEW_STEPS 单例注册表）。
- steps/planning/：plan-cycle-step.ts（PlanCycleStep class + 单例）+ shared.ts（动作解析 / 协议 /
  会话上下文 / 计划归一）+ index.ts。
- describe + review 由「并行合并一步」拆为顺序两步（describe-step → review-step），各记一条 plan 步并
  各跑一个只读工具；新增 describe / review 两条本地化步骤文案（四语言），orchestrator.test 步序断言
  相应更新为 ['plan','plan','judge','plan']。
- orchestrator / planner 退化为瘦驱动，按 index 引用注册表 / 单步。
