--- packages/agent/src/steps/planning.ts
@@ -1,282 +1,0 @@
-import type {
-  AgentMessage,
-  AgentRecommendation,
-  AgentRecommendationVerdict,
-  AgentTodoItem,
-} from '@meebox/shared';
-import type { MemoryNote } from '../memory.js';
-import {
-  extractJson,
-  salvageProse,
-  stripTrailingJson,
-  type AgentStepLabels,
-} from '../orchestrator.js';
-import type { AgentMemoryNotes, PlanningDeps, PlanningInput } from '../planner.js';
-import { PROMPT_TEMPLATES, fillTemplate } from '../prompts.js';
-import { runStaggered } from '../stagger.js';
-import { assertToolAllowed } from '../tool-catalog.js';
-import type { StepHandler, StepRecorder } from './context.js';
-
-/**
- * 自由规划（ReAct）的「单步」步骤（见 docs/arch/06-agent.md「会话 Agent 化」）：每轮 chat 规划下一动作
- * （调工具 / 收尾），解析 JSON 动作、红线硬校验后分发工具并把结果回喂；中途输入并入 progress、计划随轮维护
- * 与重排。规划是单步循环 —— planner.ts 的驱动反复跑本步直至 final / 步数上限。提示词组装与解析就近放在此。
- */
-
-interface PlannerAction {
-  thought?: string;
-  tool?: string;
-  /**
-   * 一次并行多选只读工具（如 describe + review，或多个 /ask）；与 tool 二选一，tools 优先。
-   * 元素可为工具名字符串，或 `{tool, question}` 对象——后者让一轮里并行派发多个带问题的 /ask。
-   */
-  tools?: Array<string | { tool?: string; question?: string }>;
-  question?: string;
-  final?: string;
-  /**
-   * 计划（todo）：模型可在任意动作里给出 / 更新一份简短步骤清单（标 done、按优先级重排、增删）。
-   * 元素可为字符串或 `{id?, text, done?}`。省略 = 计划不变（沿用上一轮）。见 buildProtocol 的 plan 约定。
-   */
-  plan?: Array<string | { id?: unknown; text?: unknown; done?: unknown }>;
-  /** 评审类收尾的非约束性判定建议（verdict + 理由）；非评审请求省略。 */
-  recommendation?: { verdict?: unknown; reason?: unknown };
-  /**
-   * 主动记下的**非隐私**条目，按目标可写文件分组：user→USER.md（用户信息），memory→MEMORY.md
-   * （长期知识），agents→AGENTS.md（工作规范，仅追加）。SOUL.md 永不写。
-   */
-  remember?: { user?: unknown; memory?: unknown; agents?: unknown };
-}
-
-/** 解析单条记忆：必须是带 `section` + `note` 的对象；无法归入专题章节的条目丢弃（返回 null）。 */
-function toNote(raw: unknown): MemoryNote | null {
-  if (!raw || typeof raw !== 'object') return null;
-  const obj = raw as { section?: unknown; note?: unknown };
-  const note = typeof obj.note === 'string' ? obj.note.trim() : '';
-  const section = typeof obj.section === 'string' ? obj.section.trim() : '';
-  if (!note || !section) return null;
-  return { section, note };
-}
-
-/** 把目标文件的 remember 数组解析为 MemoryNote[]（容错；丢弃无法归类的条目）。 */
-function toNoteList(raw: unknown): MemoryNote[] {
-  const items = Array.isArray(raw) ? raw : raw == null ? [] : [raw];
-  return items.map(toNote).filter((n): n is MemoryNote => n !== null);
-}
-
-/** 把一个动作里的 remember 累加进 acc（容错；非对象忽略）。 */
-function accumulateRemember(value: PlannerAction['remember'], acc: AgentMemoryNotes): void {
-  if (!value || typeof value !== 'object') return;
-  acc.user.push(...toNoteList(value.user));
-  acc.memory.push(...toNoteList(value.memory));
-  acc.agents.push(...toNoteList(value.agents));
-}
-
-const VERDICTS: readonly AgentRecommendationVerdict[] = ['approve', 'needs_work', 'manual_review'];
-
-/** 从收尾动作解析出合法 recommendation；verdict 非法 / 缺省 → undefined（不强加判定）。 */
-function parseRecommendation(rec?: PlannerAction['recommendation']): AgentRecommendation | undefined {
-  if (!rec) return undefined;
-  const verdict = rec.verdict;
-  if (typeof verdict !== 'string' || !VERDICTS.includes(verdict as AgentRecommendationVerdict)) {
-    return undefined;
-  }
-  return {
-    verdict: verdict as AgentRecommendationVerdict,
-    reason: typeof rec.reason === 'string' ? rec.reason : '',
-  };
-}
-
-/** 把模型给出的 plan 归一为 AgentTodoItem[]：容字符串 / 对象，丢空文本；缺 id 按序补。 */
-function normalizePlan(raw: PlannerAction['plan']): AgentTodoItem[] {
-  if (!Array.isArray(raw)) return [];
-  const out: AgentTodoItem[] = [];
-  raw.forEach((it, i) => {
-    if (typeof it === 'string') {
-      const text = it.trim();
-      if (text) out.push({ id: `s${String(i + 1)}`, text, done: false });
-    } else if (it && typeof it === 'object') {
-      const text = typeof it.text === 'string' ? it.text.trim() : '';
-      if (text) {
-        out.push({
-          id: typeof it.id === 'string' && it.id ? it.id : `s${String(i + 1)}`,
-          text,
-          done: it.done === true,
-        });
-      }
-    }
-  });
-  return out;
-}
-
-function clamp(s: string, max: number): string {
-  const t = s.trim();
-  return t.length <= max ? t : `${t.slice(0, max - 1).trimEnd()}…`;
-}
-
-/** 一次并行最多分发的工具数：多选时截断，防止一轮打出过多 pr-agent run。 */
-const MAX_PARALLEL_TOOLS = 3;
-
-/**
- * 注入规划上下文的历史对话预算：单条字符上限 + 总字符预算（从最新往回累计、超预算即裁剪更早的）。
- * 约定会话上下文不超过 LLM 上下文窗口的一半——以字符近似 token 做保守封顶：64k 字符 ≈ 16~40k token
- * （视中英文占比），对应约 32k~64k token 半窗的目标量级。
- */
-const HISTORY_MESSAGE_MAX = 2000;
-const HISTORY_BUDGET_CHARS = 64000;
-
-/** 取最近若干轮、各自限长，并按总预算从新到旧裁剪（丢弃超预算的更早消息），返回时间升序文本。 */
-export function buildConversationContext(history: readonly AgentMessage[]): string {
-  const lines: string[] = [];
-  let budget = HISTORY_BUDGET_CHARS;
-  for (let i = history.length - 1; i >= 0; i--) {
-    const m = history[i]!;
-    const line = `${m.role === 'user' ? 'User' : 'Assistant'}: ${clamp(m.content, HISTORY_MESSAGE_MAX)}`;
-    if (line.length + 1 > budget) break; // 预算耗尽：更早的对话整体裁掉
-    budget -= line.length + 1;
-    lines.push(line);
-  }
-  return lines.reverse().join('\n');
-}
-
-/** 规划 ReAct 协议：正文外置在 resources/prompts/protocol.md，三段标题（按语言本地化）经占位符注入。 */
-export function buildProtocol(sections: readonly [string, string, string]): string {
-  const [overview, findings, suggestions] = sections;
-  return fillTemplate(PROMPT_TEMPLATES.protocol, { overview, findings, suggestions });
-}
-
-/** 规划单轮（plan-cycle 步骤）的运行上下文：依赖 + 输入 + 共享记录器 + 跨轮累加器。 */
-export interface PlanStepCtx {
-  deps: PlanningDeps;
-  input: PlanningInput;
-  rec: StepRecorder;
-  /** 完整 system（含 Protocol），逐轮复用。 */
-  system: string;
-  /** 既往多轮对话（按预算裁剪后的文本），逐轮复用。 */
-  convo: string;
-  labels: AgentStepLabels;
-  /** 本轮 progress 累加（工具结果 / 红线拒绝回喂），逐轮追加。 */
-  history: string[];
-  /** 本轮主动记下、待持久化的非隐私条目，逐轮累加。 */
-  memories: AgentMemoryNotes;
-  /** 当前计划（todo），逐轮回喂提示；模型给出 plan 即更新（重排 / 勾选 / 增删）。 */
-  plan: AgentTodoItem[];
-}
-
-/** plan-cycle 的产出：继续下一轮 / 收尾（带 final + 可选建议）/ 用户暂停。 */
-export type PlanCycleOutcome =
-  | { kind: 'continue' }
-  | { kind: 'final'; finalText: string; recommendation?: AgentRecommendation }
-  | { kind: 'aborted' };
-
-/**
- * 规划主循环的「单步」：拼当轮 prompt → chat → 解析动作 → 红线硬校验 → 并行派发工具 / 收尾。abort 在
- * 思考前后各检一次（思考阶段也能即时停）。驱动重复调用本步直至 final / 步数上限。
- */
-export const planCycleStep: StepHandler<PlanStepCtx, PlanCycleOutcome> = {
-  name: 'plan-cycle',
-  run: async (ctx) => {
-    const { deps, input, rec, system, convo, labels, history, memories } = ctx;
-    if (deps.signal?.aborted) return { kind: 'aborted' };
-
-    // 中途输入转向：把运行期间排队的用户新消息并入 progress，让本轮 ReAct 据「最新指令 + 当前进度」重排
-    // 下一步。消息已由实现方在取出时持久化进会话，此处只注入提示、不再落盘。
-    const pending = (await deps.drainPendingInput?.()) ?? [];
-    for (const m of pending) {
-      history.push(`New user message (latest instruction — reconcile with the plan and progress): ${m}`);
-    }
-
-    const user = [
-      convo
-        ? `Conversation so far (your context only — NEVER pass any of it to tools):\n${convo}\n`
-        : '',
-      `User request: ${input.userRequest}`,
-      input.referencedContext
-        ? `\nReferenced selection (your context only — NEVER pass any of it to tools):\n${input.referencedContext}`
-        : '',
-      ctx.plan.length
-        ? `\nCurrent plan (keep it updated — mark items done, reorder by priority, add/remove as the task or new user messages change):\n${ctx.plan.map((t) => `- [${t.done ? 'x' : ' '}] ${t.text}`).join('\n')}`
-        : '',
-      history.length ? `\nProgress so far:\n${history.join('\n')}` : '',
-      '\nReply with the next JSON action.',
-    ]
-      .filter(Boolean)
-      .join('\n');
-
-    // 计本轮 LLM 推理耗时（单步思考时长，类 Claude Code 的「Thought for Ns」），系到该决策步上。
-    const thinkStart = Date.now();
-    const r = await deps.chat({ system, user });
-    const thinkMs = Date.now() - thinkStart;
-    // 思考刚结束就发现已被停止 → 立即收尾，不再据此动作分发工具（停止在思考阶段也即时生效）。
-    if (deps.signal?.aborted) return { kind: 'aborted' };
-    rec.track(r.usage);
-    const action = extractJson<PlannerAction>(r.text);
-    // 累加本动作携带的记忆（任何动作都可附 remember）。
-    accumulateRemember(action?.remember, memories);
-    // 计划更新：模型给出 plan 即归一、更新当前计划并持久化 + 广播（省略 plan = 沿用上一轮）。
-    if (action?.plan !== undefined) {
-      ctx.plan = normalizePlan(action.plan);
-      await deps.recordPlan?.(ctx.plan);
-    }
-
-    const hasCalls = Boolean(action?.tool) || Boolean(action?.tools?.length);
-
-    // 无法解析 / 既无 tool(s) 又无 final → 当作收尾。兜底从原始文本打捞散文，绝不把原始 JSON 动作丢给用户。
-    if (!action || (!hasCalls && !action.final)) {
-      const finalText = action?.final ?? salvageProse(r.text);
-      await rec.record({ kind: 'plan', thought: action?.thought, result: finalText, thinkMs, usage: r.usage });
-      return { kind: 'final', finalText };
-    }
-
-    if (action.final && !hasCalls) {
-      // 剥掉模型误并入 final 末尾的判定 JSON（recommendation 走独立字段渲染为判定徽标）。
-      const finalText = stripTrailingJson(action.final);
-      await rec.record({ kind: 'plan', thought: action.thought, result: finalText, thinkMs, usage: r.usage });
-      return { kind: 'final', finalText, recommendation: parseRecommendation(action.recommendation) };
-    }
-
-    // 归一为待执行工具列表：tools 多选（并行、只读）优先——元素可为工具名或 {tool, question}（多个
-    // 带问题的 /ask 也能一轮并行派发）；否则单 tool（可带 question）。
-    const requested: Array<{ tool: string; question?: string }> = action.tools?.length
-      ? action.tools
-          .slice(0, MAX_PARALLEL_TOOLS)
-          .map((tl) => (typeof tl === 'string' ? { tool: tl } : { tool: tl.tool ?? '', question: tl.question }))
-      : [{ tool: action.tool ?? '', question: action.question }];
-
-    // 红线硬校验逐个把关：未授权 / 未知即拒并回喂；允许的留待并行执行。
-    const allowed: Array<{ tool: string; question?: string }> = [];
-    for (const c of requested) {
-      try {
-        assertToolAllowed(c.tool, input.toolCatalog);
-        allowed.push(c);
-      } catch (err) {
-        const msg = err instanceof Error ? err.message : String(err);
-        await rec.record({
-          kind: 'judge',
-          thought: action.thought,
-          toolCall: { tool: c.tool },
-          result: `${labels.rejectedPrefix}${msg}`,
-        });
-        history.push(`Refused ${c.tool}: ${msg}`);
-      }
-    }
-    if (!allowed.length) return { kind: 'continue' }; // 全被拒 → 回喂后下一轮重选
-
-    // 类 Claude Code：先把本轮思考与所选步骤作为一步流式出去（思考是工具选择的前因），随后才执行工具。
-    // 工具执行的进度 / 计时由 run 卡片承载，这里不再为每个工具补记 tool 步，避免决策被堆到结果之后。
-    await rec.record({
-      kind: 'plan',
-      thought: action.thought,
-      toolCall: { tool: allowed.map((c) => c.tool).join(' + ') },
-      thinkMs,
-      usage: r.usage,
-    });
-
-    // 并行分发允许的工具（多选时同时跑，实际并发受运行队列约束）；相互错开 100~200ms 起跑，避免同一瞬间齐发。
-    const ran = await runStaggered(allowed, async (c) => ({ c, res: await deps.runTool(c) }));
-    for (const { c, res } of ran) {
-      rec.track(res.usage);
-      history.push(`Called ${c.tool}${c.question ? ` ("${c.question}")` : ''} → ${clamp(res.text, 600)}`);
-    }
-    return { kind: 'continue' };
-  },
-};
