+++ packages/agent/src/steps/planning/index.ts
@@ -1,0 +1,3 @@
+export { buildConversationContext, buildProtocol } from './shared.js';
+export type { PlanCycleOutcome, PlanStepCtx } from './shared.js';
+export { planCycleStep, PlanCycleStep } from './plan-cycle-step.js';
